src/requirements.txt
distro==1.5.0
numpy==1.19.4
pandas==1.2.0
python-dateutil==2.8.1
pytz==2020.5
six==1.15.0
tabula-py==2.2.0
